requirements.txt
apipkg==1.5
attrs==19.3.0
awscli==1.17.9
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
contextlib2==0.6.0.post1
cycler==0.10.0
docutils==0.15.2
execnet==1.7.1
Flask==1.1.1
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mock==3.0.5
more-itertools==8.2.0
numpy==1.18.1
packaging==20.1
paho-mqtt==1.5.0
path==13.1.0
path.py==12.4.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pyparsing==2.4.6
pytest==5.3.5
pytest-shutil==1.7.0
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.2
six==1.14.0
speedtest-cli==2.1.2
termcolor==1.1.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
zipp==2.1.0
